Prevent redis backup file being commited
`dump.rdb` is generated by redis as it's default persistance method,
and will be written to the directory redis was started in.

In a proper dev environment redis should be running as a background
service, but to get this running quickly I was running redis in
the directory of the repo - not ideal, but not uncommon.

Ensure the file is ignored, just in case.

More details: https://redis.io/topics/persistence