Store email address during user registration
`User.email` is modelled, so we can store the `email` provided by the
OmniAuth response from Github (with the `user:email` scope), see:
- https://github.com/omniauth/omniauth/wiki/Auth-Hash-Schema
- https://developer.github.com/v3/oauth/#scopes

Previous OmniAuth used the default scope, this upgrades to the 'user:email'
scope to ensure we get a users email address, and they are aware that
the app is likely to do something with an email address, like email
them. So they can make an informed decision about authing the app.

For new users this is straight forward. Any existing users in
production will have `nil` value for the new field, but can be back
filled when logging in again, as the oAuth callback will contain the
same data (see link above).

This would allow `email` (and also `name`) to be updated on each login,
if they've changed. This change is to enable sending email notification
on like/hate, recording email for new users (and backfilling for existing)
is enough to do that. Behaviour like updating can be looked at later.

I've also added some testing around `UserRegistration` to make the intent
clearer, and future changes easier.