Add Mailer to tell a user someone voted on their movie
Create a `UserMailer`, with a `movie_vote` mail. It's likely we might
add other emails to users in the future, not just votes.

Add a simple plain text mail template that contains the required
information, with testing to assert it's being rendered correctly.

Sets a `from` address for different environments - but the production
one needs updating (and this comment to be removed) once a real value
has been agreed.

This creates the framework code requires to render an email, but
doesn't have strong opinions on the content, which can be iterated.

Notes on content:

The content is quite MVP, and needs some improvement based on what
the user need of this feature is. With a toy project it's more
difficult to guess at why the user would want this email, possibly
to find out more about the user who liked/hated the movie, to see
the overall like/hate count - all of which could be put into the
email itself. If the email is more of a business need - driving
traffic back to the site - then you'd omit some of the information
mentioned before, to require them to visit the site again (but
the site doesn't say who the likes/hates are by).

It's also not clear where we'd want to direct a user, as the site
doesn't have a permalink to a movie or a user page, to find out
more about the user.

Because of this i've left the content intentionally simple/vague,
but in the real world would talk to a combination of a Product
Manager, Interaction Designer and Content designer to get a better
idea of what we want to email to achieve.