Trigger an email when someone votes on a movie
Add a spec for `VotingBooth` to ensure `UserMailer.movie_vote` is
called with the required arguments - but not the implementation
of the Mailer itself - and that it returns a object with a `deliver`
method.

I'm not entirely happy with the spec implementation. I'm not super
familiar with `rspec`, and although this works, it feels like it
might not be doing things in the rspec way. Might benefit from a
refactor later.