Extract mail delivery settings into ENV vars
Depending on how the app will send email in production (Amazon SES,
Sendgrid, internal SMTP, etc) this configuration will need to change.

It may need additional authentication configuration, depending on the
chosen delivery mechanism. It could be as simple as a host/port on
the local network, a remote SMTP provider, or even local sendmail.

I've chosen `address` and `port` to illustrate the pattern, but this
will need to be updated before going into production.

More detail on the mail configuration options, including example
setups, here:

http://guides.rubyonrails.org/action_mailer_basics.html#action-mailer-configuration